requirements.txt
gymnasium~=0.28.1
pytesseract~=0.3.10
matplotlib~=3.7.2
fuzzywuzzy~=0.18.0
pillow~=10.2.0
pyautogui~=0.9.54
keyboard~=0.13.5
psutil~=5.9.6
pyyaml~=6.0.1
numpy~=1.26.0
pydirectinput~=1.0.4
pygetwindow~=0.0.9
pandas~=2.1.4
pymem~=1.13.0
pywin32~=306